secret/requirements.txt
azureml-core==1.50.0
azureml-dataset-runtime==1.50.0